+++ app/serializers/grade_serializer_v2.rb
@@ -1,6 +1,5 @@
 class GradeSerializerV2 < ActiveModel::Serializer
-  attribute :lesson_id, key: :lesson_id
-  attributes :student_id, :deleted_at, :skill_id, :mark
+  attributes :lesson_id, :student_id, :deleted_at, :skill_id, :mark
 
   belongs_to :student
   belongs_to :lesson